Move redis stuff into game
